Separate consumerInstance into a new file
